+++ src/models/Query.js
@@ -106,3 +106,6 @@ class Query {
 }
 
 export default Query;
+export {
+  ApiResponseError,
+};
